+++ pkg/deployment/resources/pod_inspector.go
@@ -248,7 +248,7 @@ func (r *Resources) InspectPods(ctx context.Context, cachedStatus inspector.Insp
 					case api.MemberPhaseNone:
 						// Do nothing
 						log.Debug().Str("pod-name", podName).Msg("PodPhase is None, waiting for the pod to be recreated")
-					case api.MemberPhaseShuttingDown, api.MemberPhaseUpgrading, api.MemberPhaseFailed:
+					case api.MemberPhaseShuttingDown, api.MemberPhaseUpgrading, api.MemberPhaseFailed, api.MemberPhaseRotateStart:
 						// Shutdown was intended, so not need to do anything here.
 						// Just mark terminated
 						wasTerminated := m.Conditions.IsTrue(api.ConditionTypeTerminated)